subscribeToMatrix3D functions: Reality Editor 1.5.8
This is a new function in the Reality Editor where your interface can
subscribe to the 3D model View 4x4 Transformation Matrix. Short: Use
getPossitionX(), getPossitionY(), getPossitionZ() to get the x, y and z
position of your object relative to the camera.

use update() to update your interface synchronized with the video image